+++ MK4duo/src/motion/delta_parameters.h
@@ -43,7 +43,8 @@
             base_min_pos[ABC]         = { X_MIN_POS, Y_MIN_POS, Z_MIN_POS },
             base_max_pos[ABC]         = { X_MAX_POS, Y_MAX_POS, Z_MAX_POS },
             base_home_pos[ABC]        = { X_HOME_POS, Y_HOME_POS, Z_HOME_POS },
-            max_length[ABC]           = { X_MAX_LENGTH, Y_MAX_LENGTH, Z_MAX_LENGTH };
+            max_length[ABC]           = { X_MAX_LENGTH, Y_MAX_LENGTH, Z_MAX_LENGTH },
+            clip_start_height         = Z_MAX_POS;
 
       // Function
       void Init();
@@ -57,6 +58,7 @@
 
       void convertIncomingEndstops();
       void convertOutgoingEndstops();
+      void Set_clip_start_height();
 
     private: